Refactored the LTIWorker. Hydra is now functional again. Aggregation settings handling seems off still.

+++ src/hypro/datastructures/reachability/Settings.h
@@ -20,7 +20,7 @@ struct AnalysisParameters {
 
 struct Settings {
 	std::vector<AnalysisParameters> strategy{ 1 };
-	int jumpDepth{ std::numeric_limits<int>::max() };
+	size_t jumpDepth{ std::numeric_limits<int>::max() };
 	tNumber localTimeHorizon{ tNumber( 0 ) };
 	std::vector<std::vector<std::size_t>> plotDimensions{ { 0, 1 } };
 	std::vector<std::string> plotFileNames{ "out" };
